requirements.txt
colorama==0.3.7
Django==1.8.14
djangorestframework==3.4.5
httpie==0.9.6
Pygments==2.1.3
requests==2.11.1
